OverTheWire Write-ups
This commit involves the first ~15 levels of Bandit
which I never took the time to write up the first time
I completed the challenges.

Also, I copied over the full write up for Leviathan but it
is in massive need of a formatting overhaul.

Narnia has one? exercise done. Doesn't matter it's been copied